+++ apps/frontend-web/src/components/ProjectTabBar.tsx
@@ -7,7 +7,6 @@ import { SortableProjectTab } from './SortableProjectTab';
 import { ProjectSelector } from './settings/ProjectSelector';
 import { ClaudeCodeStatusBadge } from './ClaudeCodeStatusBadge';
 import { CLIToolStatusBadge } from './CLIToolStatusBadge';
-import { LocalLLMStatusBadge } from './LocalLLMStatusBadge';
 import { OpenAIEndpointsStatusBadge } from './OpenAIEndpointsStatusBadge';
 import { useProjectStore } from '../stores/project-store';
 import { useSettingsStore, saveSettings } from '../stores/settings-store';
@@ -160,7 +159,6 @@ export function ProjectTabBar({
         <Separator orientation="vertical" className="h-4 mx-0.5" />
         <ClaudeCodeStatusBadge iconOnly onOpenOnboarding={onOpenOnboarding} />
         <CLIToolStatusBadge iconOnly />
-        <LocalLLMStatusBadge iconOnly />
         <OpenAIEndpointsStatusBadge iconOnly />
         <Separator orientation="vertical" className="h-4 mx-0.5" />
         <Button